Major Update #2
Major update to the code of the page since last commit. Some highlights
include changes to the structure and styling of the page, addition of
options that allow user to symbolize pay stations based on data for
particular times, multi-threading for loading data onto page, and more
odds and ends.